Version 6.4.0: Re-Write of the Marlin Module

+++ Marlin/HTTPMessage.cpp
@@ -606,7 +606,7 @@ HTTPMessage::SetFile(CString& p_fileName)
 CString
 HTTPMessage::GetRoute(int p_index)
 {
-  if(p_index >= 0 && p_index < m_routing.size())
+  if(p_index >= 0 && p_index < (int)m_routing.size())
   {
     return m_routing[p_index];
   }
